Move React.Flux.DOM to React.DOM as it does not depend on flux
