+++ react-hs/src/React/Flux/Outdated.hs
@@ -49,10 +49,10 @@ module React.Flux.Outdated
 import Control.Monad.Writer
 import Data.Typeable
 
+import React.DOM (div_)
 import React.Flux.Store
 import React.Flux.Internal
 import React.Flux.View (ViewEventHandler, StatefulViewEventHandler)
-import React.Flux.DOM (div_)
 
 import System.IO.Unsafe (unsafePerformIO)
 import JavaScript.Array
